rename classes improving coherence with Qt naming
