+++ CMakeLists.txt
@@ -37,10 +37,10 @@ endif()
 add_library(
     ${PROJECT_NAME}
 
-    include/${PROJECT_NAME}/tags.hpp
-    include/${PROJECT_NAME}/tags_area.hpp
-    src/tags.cpp
-    src/tags_area.cpp
+    include/${PROJECT_NAME}/tags_line_edit.hpp
+    include/${PROJECT_NAME}/tags_edit.hpp
+    src/tags_line_edit.cpp
+    src/tags_edit.cpp
 )
 
 set_target_properties(${PROJECT_NAME} PROPERTIES CXX_STANDARD 17 AUTOMOC ON)
